fix: add daily aggregate sync to foreground and improve cumulative metric fallback
Steps and other cumulative metrics stopped appearing because daily aggregates
were only sent via the background SyncWorker. If that path fails silently,
the frontend shows 0.

Changes:
- Extract daily aggregate fetch/send into shared DailyAggregateSync.kt
- Add daily aggregate sync to foreground syncNow() in MainActivity
- Add backend logging when daily-aggregates endpoint receives data
- Add getRawDailySum fallback for goals when no aggregates exist
- Improved diagnostic logging for zero/null aggregate values
