+++ apps/backend/src/db/index.ts
@@ -67,6 +67,7 @@ export {
   deleteTimeSeriesMetric,
   deleteTimeSeriesPoint,
   getDailyAggregates,
+  getRawDailySum,
   getTimeSeries,
   getTimeSeriesBucketed,
   getTimeSeriesMultiMetric,
