+++ apps/backend/src/sync-router.ts
@@ -128,6 +128,13 @@ export const createSyncRouter = (deps: SyncRouterDeps, authMiddleware: RequestHa
       const { data } = req.body
       const user = req.user!
 
+      const metrics = [...new Set(data.map((a) => a.metric))]
+      const dates = [...new Set(data.map((a) => a.date))].sort()
+      console.log(
+        `📊 Daily aggregates received: ${data.length} entries for [${metrics.join(', ')}] ` +
+          `dates ${dates[0]}..${dates[dates.length - 1]}`,
+      )
+
       for (const aggregate of data) {
         await deps.processDailyAggregate(user, aggregate)
       }
